Added the delete button
I now have the cards showing in the 'library', the add book button
in the sidebar showing a form to add a new book. after adding a new
book, the library gets refreshed to show the new book.

And with this update, there is an 'X' on each card that will delete
the card if it is clicked and delete from the database and refresh
the library on screen.