fixed the bug that read / not read was not correct
when I added a new book, it would always say 'read'.
i was just getting the fourth value from 'forms' which was
just the value of the first radio button (not the one that was
checked). so, had to find all the radios and find the one
that was checked and then get its value.  Then....had to
translate that value to boolean from string.